+++ src/fs_poll.c
@@ -21,9 +21,10 @@
 #include <stddef.h>
 #include <uv.h>
 
+#include "fs.h"
 #include "handle.h"
+#include "internal.h"
 #include "luv.h"
-#include "private.h"
 
 LUV_DEFAPI luaL_Reg luv_fs_poll_methods[] = {
   {"start", luv_fs_poll_start},
@@ -114,7 +115,7 @@ LUV_LUAAPI int luv_fs_poll_stop(lua_State *L) {
 LUV_LUAAPI int luv_fs_poll_getpath(lua_State *const L) {
   uv_fs_poll_t *fs_poll = luv_check_fs_poll(L, 1);
 
-  size_t len = 2 * PATH_MAX;
+  size_t len = 2 * (size_t)PATH_MAX;
   char buf[2 * PATH_MAX];
 
   const int ret = uv_fs_poll_getpath(fs_poll, buf, &len);
